requirements.txt
PyQt5==5.15.11
requests==2.32.5
Pillow==12.0.0
py-machineid==0.8.1
curl_cffi==0.13.0
zendriver==0.14.2
qasync==0.28.0

# Video Analyzer Tool Dependencies
opencv-python==4.10.0.84
scenedetect[opencv]==0.6.4
yt-dlp==2024.12.23
anthropic==0.39.0
tqdm==4.67.1
python-dotenv==1.0.1